feat: Add tests for BlobContainerService and StorageAccountService
- Changed reconciliation loop for BlobContainerReconciler and FileShareReconciler to set storage account name in annotations from workflow context.
- Blob storage lifetime not set if no time provided

+++ src/main/java/no/fintlabs/azure/storage/blob/BlobContainerSpec.java
@@ -9,5 +9,5 @@ import no.fintlabs.azure.AzureSpec;
 @Getter
 public class BlobContainerSpec implements AzureSpec {
 
-    private Long lifespanDays;
+    private Integer lifespanDays;
 }
